+++ src/theme/MDXComponents.js
@@ -4,7 +4,7 @@ import PrettyLink from '@site/src/components/PrettyLink'
 import ContractData from '@site/src/components/ContractData'
 import ContractAddress from '@site/src/components/StaticContractAddress'
 import AddressCheck from '@site/src/components/AddressCheck'
-import VeYFICalculator from '../components/veYFI-calc'
+import VeYFICalculator from '../components/veYFI-calculator'
 
 /**
  * Manually add the custom components to the list of MDXComponents that docusaurus uses
